Reorganize input sources
